fix modal connections for submits
