+++ client/src/components/common/TimeRangeSlider/index.js
@@ -1,89 +1,139 @@
-import React, { Component } from 'react'
-import InputRange from 'react-input-range'
-import './styles.css'
-import { minuteToTime, timeToMinute } from '../../../utils'
-import moment from 'moment'
+import React from 'react'
+import styled from '@emotion/styled'
+import system from '../../../design/theme'
+import Zoom from 'react-reveal'
+import Checkbox from './Checkbox'
+import TimeRangeSlider from './Slider'
+import Status from './Status'
 import PropTypes from 'prop-types'
 
-const formatHours = minutes => {
-  let clearMinutes = minutes < 0 ? 0 : minutes
-  const time = minuteToTime(clearMinutes, 24)
-  return moment({ ...time }).format('h:mm a')
+const Slider = ({
+  id,
+  name,
+  closedAllDay,
+  status,
+  toggled,
+  disabled,
+  start,
+  draggableTrack,
+  onChange,
+  end
+}) => {
+  return (
+    <Container>
+      <div>
+        <p className="days" id={id}>
+          {name}
+        </p>
+      </div>
+      <Zoom left>
+        <div className="closeToggle">
+          <Checkbox onToggle={closedAllDay} toggled={toggled} name={name} />
+          <Status status={status} />
+        </div>
+      </Zoom>
+      <Zoom left>
+        <div className="buttons">
+          <TimeRangeSlider
+            draggableTrack={draggableTrack}
+            onChange={onChange}
+            disabled={disabled}
+            name={name}
+            start={start}
+            end={end}
+          />
+          <div className="row">
+            <p>{start}</p>
+            <p>{end}</p>
+          </div>
+        </div>
+      </Zoom>
+    </Container>
+  )
 }
 
-class TimeRangeSlider extends Component {
-  constructor(props) {
-    super(props)
-    this.state = {}
-  }
+export default Slider
 
-  onChange = ({ min, max }) => {
-    this.props.onChange({
-      start: formatHours(min),
-      end: formatHours(max)
-    })
-  }
-
-  onChangeComplete = ({ min, max }) => {
-    this.props.onChangeComplete({
-      start: formatHours(min),
-      end: formatHours(max)
-    })
-  }
-
-  onChangeStart = ({ min, max }) => {
-    this.props.onChangeStart({
-      start: formatHours(min),
-      end: formatHours(max)
-    })
-  }
-
-  render() {
-    let { start, end } = this.props,
-      min = timeToMinute(start),
-      max = timeToMinute(end)
-
-    return (
-      <InputRange
-        disabled={this.props.disabled}
-        draggableTrack={this.props.draggableTrack}
-        minValue={timeToMinute(this.props.minValue)}
-        maxValue={timeToMinute(this.props.maxValue)}
-        onChangeStart={this.onChangeStart}
-        onChange={this.onChange}
-        onChangeComplete={this.onChangeComplete}
-        step={15}
-        value={{ min, max }}
-        name={this.props.name}
-      />
-    )
-  }
-}
-
-TimeRangeSlider.defaultProps = {
-  disabled: false,
-  draggableTrack: true,
-  format: 12,
-  maxValue: '11:59pm',
-  minValue: '12:00am',
-  onChange: () => {},
-  onChangeComplete: () => {},
-  onChangeStart: () => {},
-  step: 15
-}
-
-export default TimeRangeSlider
-
-TimeRangeSlider.propTypes = {
-  disabled: PropTypes.bool.isRequired,
-  draggableTrack: PropTypes.bool.isRequired,
-  minValue: PropTypes.string.isRequired,
-  maxValue: PropTypes.string.isRequired,
-  onChangeStart: PropTypes.func.isRequired,
-  onChange: PropTypes.func.isRequired,
-  onChangeComplete: PropTypes.func.isRequired,
-  format: PropTypes.number.isRequired,
+Slider.propTypes = {
+  id: PropTypes.number.isRequired,
   name: PropTypes.string.isRequired,
-  value: PropTypes.object,
-  step: PropTypes.number.isRequired
+  closedAllDay: PropTypes.func.isRequired,
+  onChange: PropTypes.func.isRequired,
+  toggled: PropTypes.bool.isRequired,
+  disabled: PropTypes.bool.isRequired,
+  start: PropTypes.string.isRequired,
+  end: PropTypes.string.isRequired
 }
+
+const Container = styled('div')`
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+  justify-content: center;
+  margin: 0 auto;
+  margin-bottom: 7.5px;
+  padding-bottom: 7.5px;
+  border-bottom: 2px solid ${system.color.neutral};
+
+  :last-of-type {
+    border-bottom: none;
+  }
+
+  .days {
+    border: 0;
+    padding: ${system.spacing.standardPadding};
+    width: 116px;
+    text-align: left;
+    margin: 0 10px 10px;
+    font-size: ${system.fontSizing.sm};
+    font-weight: bold;
+  }
+  .buttons {
+    width: 500px;
+    background: white;
+    padding: 12px;
+    margin-right: 15px;
+    border-radius: 0 25px 25px 0;
+    @media ${system.breakpoints[1]} {
+      width: 100%;
+      border-radius: 0;
+    }
+
+    button {
+      cursor: pointer;
+      margin: 5px;
+      font-size: 1.8rem;
+      border: transparent;
+    }
+    .row {
+      display: flex;
+      flex-flow: row nowrap;
+      justify-content: space-between;
+      align-items: center;
+
+      p {
+        font-family: 'Nunito', sans-serif;
+        margin-top: 5px;
+      }
+    }
+  }
+  .closeToggle {
+    display: flex;
+    flex-direction: column;
+    justify-content: space-evenly;
+    height: 6px;
+    position: relative;
+    margin-bottom: 12px;
+    padding: 0 15px;
+    cursor: pointer;
+    -webkit-user-select: none;
+    -moz-user-select: none;
+    -ms-user-select: none;
+    user-select: none;
+    p {
+      margin-top: 5px;
+      font-size: ${system.fontSizing.s};
+      height: 100%;
+    }
+  }
+`
